# brag-output/work/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: compliance-engine API
  version: 1.0.0
  description: >-
    Versioned REST API of the compliance engine. Every authenticated request is
    scoped to the tenant workspace bound to its bearer token. Outputs are
    operational readiness aids, not legal advice and not certification.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/about:
    get:
      tags: [service]
      operationId: about
      x-permission: data.read
      summary: Edition, version, supported schemas and loaded catalogs
      responses:
        "200":
          description: Server description for the calling workspace.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/About"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/adapters/{name}/manifest:
    put:
      tags: [ingestion]
      operationId: putManifest
      x-permission: data.write
      summary: Register or replace an adapter capability manifest
      parameters:
        - {name: name, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Manifest"}
      responses:
        "200":
          description: The registered manifest.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Manifest"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "422": {$ref: "#/components/responses/Error"}
  /api/v1/ingestions:
    post:
      tags: [ingestion]
      operationId: postIngestion
      x-permission: data.write
      summary: Ingest a canonical batch (or validate it with dry_run=true)
      parameters:
        - {name: dry_run, in: query, required: false, schema: {type: boolean, default: false}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Batch"}
      responses:
        "200":
          description: No changes (identical content), or a dry-run result with the completeness of the would-be snapshot.
          content:
            application/json:
              schema:
                oneOf:
                  - {$ref: "#/components/schemas/IngestionResult"}
                  - {$ref: "#/components/schemas/DryRunResult"}
        "201":
          description: A new snapshot was committed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IngestionResult"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "409": {$ref: "#/components/responses/Error"}
        "413": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /api/v1/ingestions/{id}:
    get:
      tags: [ingestion]
      operationId: getIngestion
      x-permission: data.read
      summary: Stored record of an ingestion or rollback
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: The ingestion.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Ingestion"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/ingestions/{id}/rollback:
    post:
      tags: [ingestion]
      operationId: rollbackIngestion
      x-permission: data.write
      summary: Restore the snapshot preceding an ingestion as a new snapshot
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "201":
          description: The rollback commit.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RollbackResult"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/snapshots:
    get:
      tags: [snapshots]
      operationId: listSnapshots
      x-permission: data.read
      summary: Every snapshot (revision) of the workspace, oldest first
      responses:
        "200":
          description: Revisions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items: {$ref: "#/components/schemas/Revision"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/snapshots/{id}:
    get:
      tags: [snapshots]
      operationId: getSnapshot
      x-permission: data.read
      summary: Snapshot summary with record counts per entity
      parameters:
        - {$ref: "#/components/parameters/SnapshotID"}
      responses:
        "200":
          description: Snapshot summary.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SnapshotSummary"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/snapshots/{id}/records/{entity}:
    get:
      tags: [snapshots]
      operationId: getRecords
      x-permission: data.read
      summary: Records of one canonical entity with per-field states
      parameters:
        - {$ref: "#/components/parameters/SnapshotID"}
        - {name: entity, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: Records.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RecordsResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/snapshots/{id}/completeness:
    get:
      tags: [snapshots]
      operationId: getCompleteness
      x-permission: data.read
      summary: Data-completeness view (register gaps, codelists, dangling references)
      parameters:
        - {$ref: "#/components/parameters/SnapshotID"}
      responses:
        "200":
          description: Completeness view.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Completeness"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/provenance:
    get:
      tags: [snapshots]
      operationId: getProvenance
      x-permission: data.read
      summary: Change history of one record
      parameters:
        - {name: entity, in: query, required: true, schema: {type: string}}
        - {name: key, in: query, required: true, description: 'Record identity key, e.g. ["SAMPLETP000000000003"]', schema: {type: string}}
      responses:
        "200":
          description: Provenance entries, oldest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: {$ref: "#/components/schemas/ProvenanceEntry"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/catalogs:
    get:
      tags: [catalogs]
      operationId: listCatalogs
      x-permission: data.read
      summary: Loaded control catalog versions
      responses:
        "200":
          description: Catalog summaries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  catalogs:
                    type: array
                    items: {$ref: "#/components/schemas/CatalogSummary"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/catalogs/diff:
    get:
      tags: [catalogs]
      operationId: diffCatalogs
      x-permission: data.read
      summary: Differences between two catalog versions
      parameters:
        - {name: a, in: query, required: true, description: catalog@version, schema: {type: string}}
        - {name: b, in: query, required: true, description: catalog@version, schema: {type: string}}
      responses:
        "200":
          description: The diff.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CatalogDiff"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/catalogs/{catalog}/{version}:
    get:
      tags: [catalogs]
      operationId: getCatalog
      x-permission: data.read
      summary: One catalog version with its controls
      parameters:
        - {name: catalog, in: path, required: true, schema: {type: string}}
        - {name: version, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: The catalog.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Catalog"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/evaluations:
    post:
      tags: [evaluations]
      operationId: postEvaluation
      x-permission: evaluation.run
      summary: Evaluate a snapshot against catalog versions and store the result
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: "#/components/schemas/EvaluationRequest"}
      responses:
        "201":
          description: The stored evaluation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Evaluation"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/evaluations/{id}:
    get:
      tags: [evaluations]
      operationId: getEvaluation
      x-permission: data.read
      summary: A stored evaluation
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "200":
          description: The evaluation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Evaluation"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/imports:
    post:
      tags: [ingestion]
      operationId: postImport
      x-permission: data.write
      summary: Import a CSV or XLSX file (validate only with dry_run=true)
      parameters:
        - name: dry_run
          in: query
          required: false
          schema: {type: boolean, default: false}
        - name: mode
          in: query
          required: false
          schema: {type: string, enum: [full, incremental], default: incremental}
        - name: entity
          in: query
          required: false
          description: Canonical entity of a CSV file. Default is the file name without its extension.
          schema: {type: string}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: A <entity>.csv file or a .xlsx workbook built from the template.
      responses:
        "200":
          description: Dry-run report, or no changes (identical content).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportResult"}
        "201":
          description: A new snapshot was committed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ImportResult"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "409": {$ref: "#/components/responses/Error"}
        "413": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /api/v1/templates.xlsx:
    get:
      tags: [ingestion]
      operationId: getTemplateWorkbook
      x-permission: data.read
      summary: Import workbook template with a README sheet and one sheet per entity
      responses:
        "200":
          description: The workbook.
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: {type: string, format: binary}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/templates/{file}:
    get:
      tags: [ingestion]
      operationId: getTemplateCSV
      x-permission: data.read
      summary: Header-only CSV import template of one entity
      parameters:
        - name: file
          in: path
          required: true
          description: <entity>.csv
          schema: {type: string}
      responses:
        "200":
          description: The CSV template.
          content:
            text/csv:
              schema: {type: string}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/me:
    get:
      tags: [access]
      operationId: getMe
      x-permission: data.read
      summary: The calling principal, its roles and permissions
      responses:
        "200":
          description: The caller.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Me"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
  /api/v1/members:
    get:
      tags: [access]
      operationId: listMembers
      x-permission: members.manage
      summary: Members of the workspace and their roles
      responses:
        "200":
          description: Members ordered by email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  members: {type: array, items: {$ref: "#/components/schemas/Member"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/members/{email}:
    put:
      tags: [access]
      operationId: putMember
      parameters:
        - {name: email, in: path, required: true, schema: {type: string, format: email}}
      x-permission: members.manage
      summary: Grant roles to a user in the workspace (creates a user without a password when needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              additionalProperties: false
              properties:
                roles: {type: array, minItems: 1, items: {$ref: "#/components/schemas/Role"}}
      responses:
        "200":
          description: The membership.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Member"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
    delete:
      tags: [access]
      operationId: deleteMember
      x-permission: members.manage
      summary: Remove a member and revoke their tokens in the workspace
      parameters:
        - {name: email, in: path, required: true, schema: {type: string, format: email}}
      responses:
        "204": {description: Removed.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/tokens:
    get:
      tags: [access]
      operationId: listTokens
      x-permission: tokens.own
      summary: API tokens of the workspace (admins) or the caller's own tokens
      responses:
        "200":
          description: Tokens without their values or hashes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens: {type: array, items: {$ref: "#/components/schemas/Token"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
    post:
      tags: [access]
      operationId: postToken
      x-permission: tokens.own
      summary: Create an API token; its value is returned once
      description: >-
        A user token acts as the caller and may carry only roles the caller holds.
        Service tokens (service=true) have no user, may hold only owner and auditor,
        and require the admin role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, roles]
              additionalProperties: false
              properties:
                name: {type: string, minLength: 1, maxLength: 100}
                roles: {type: array, minItems: 1, items: {$ref: "#/components/schemas/Role"}}
                expires_at: {type: string, format: date-time}
                service: {type: boolean}
      responses:
        "201":
          description: The token and its value. The value is never shown again.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: {$ref: "#/components/schemas/Token"}
                  value: {type: string}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/tokens/{id}:
    delete:
      tags: [access]
      operationId: deleteToken
      x-permission: tokens.own
      summary: Revoke a token (admins any token of the workspace, others their own)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "204": {description: Revoked.}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/RateLimited"}
        "501": {$ref: "#/components/responses/Error"}
  /api/v1/audit:
    get:
      tags: [access]
      operationId: getAudit
      x-permission: audit.read
      summary: The workspace's append-only audit log, optionally with chain verification
      parameters:
        - {name: after_seq, in: query, schema: {type: integer, minimum: 0}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 100}}
        - {name: verify, in: query, schema: {type: boolean}}
      responses:
        "200":
          description: Events in ascending order.
          content:
            application/json:
              schema:
                type: object
                properties:
                  events: {type: array, items: {$ref: "#/components/schemas/AuditEvent"}}
                  next_after_seq: {type: integer}
                  verification: {$ref: "#/components/schemas/AuditVerification"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "429": {$ref: "#/components/responses/RateLimited"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        API token bound to one tenant workspace and a set of roles: a static token
        configured in COMPLIANCE_TOKENS ('compliance-engine token generate'), or a
        stored token created through POST /api/v1/tokens or 'compliance-engine token create'.
        Every operation declares the permission it requires in x-permission.
...
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        WWW-Authenticate: {schema: {type: string}}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: >-
        The caller's roles lack the operation's permission (code forbidden or
        role_not_held), or the feature is not entitled (code feature_not_entitled).
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    RateLimited:
      description: Too many requests from this caller, or too many failed authentications from this address.
      headers:
        Retry-After: {schema: {type: integer}, description: Seconds to wait.}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
...
    About:
      type: object
      properties:
        edition: {type: string}
        version: {type: string}
        schema_versions: {type: array, items: {type: string}}
        catalogs: {type: array, items: {$ref: "#/components/schemas/CatalogRef"}}
        tenant_id: {type: string}
        workspace_id: {type: string}
        encryption_at_rest: {type: boolean, description: Whether sensitive values are encrypted at rest (COMPLIANCE_ENCRYPTION_KEY_FILE).}
...